Move swap-domain services to appropriate folder
