+++ src/cow-react/swap/hooks/useSwapButtonContext.ts
@@ -10,8 +10,8 @@ import { TradeType } from '@uniswap/sdk-core'
 import { computeSlippageAdjustedAmounts } from 'utils/prices'
 import { useHasEnoughWrappedBalanceForSwap, useWrapType, useWrapUnwrapError } from 'hooks/useWrapCallback'
 import { useCallback } from 'react'
-import { logSwapFlow } from 'pages/Swap/swapFlow/logger'
-import { swapFlow } from 'pages/Swap/swapFlow'
+import { logSwapFlow } from 'cow-react/swap/services/swapFlow/logger'
+import { swapFlow } from 'cow-react/swap/services/swapFlow'
 import { useApproveCallbackFromTrade } from 'hooks/useApproveCallback'
 import { useGnosisSafeInfo } from 'hooks/useGnosisSafeInfo'
 import { useIsSwapUnsupported } from 'hooks/useIsSwapUnsupported'
@@ -21,7 +21,7 @@ import { SwapButtonContext } from 'pages/Swap/components/SwapButton/SwapButton'
 import { useGetQuoteAndStatus } from 'state/price/hooks'
 import { OperationType } from 'components/TransactionConfirmationModal'
 import { useTransactionConfirmModal } from 'cow-react/swap/hooks/useTransactionConfirmModal'
-import { useSwapFlowContext } from 'pages/Swap/swapFlow/useSwapFlowContext'
+import { useSwapFlowContext } from 'cow-react/swap/hooks/useSwapFlowContext'
 import { PriceImpact } from 'hooks/usePriceImpact'
 
 export interface SwapButtonInput {
