+++ src/cow-react/swap/hooks/useSwapFlowContext.ts
@@ -1,4 +1,4 @@
-import { SwapFlowContext } from 'pages/Swap/swapFlow/types'
+import { SwapFlowContext } from 'cow-react/swap/services/swapFlow/types'
 import { useWeb3React } from '@web3-react/core'
 import { useSwapState } from 'state/swap/hooks'
 import { useDerivedSwapInfo } from 'state/swap/hooks'
@@ -9,9 +9,9 @@ import { useAddPendingOrder } from 'state/orders/hooks'
 import { useAppData } from 'hooks/useAppData'
 import { useDispatch } from 'react-redux'
 import { AppDispatch } from 'state'
-import { SwapFlowAnalyticsContext } from 'pages/Swap/swapFlow/steps/analytics'
+import { SwapFlowAnalyticsContext } from 'cow-react/swap/services/swapFlow/steps/analytics'
 import useENSAddress from 'hooks/useENSAddress'
-import { useSwapConfirmManager } from '@src/cow-react/swap/hooks/useSwapConfirmManager'
+import { useSwapConfirmManager } from 'cow-react/swap/hooks/useSwapConfirmManager'
 import { useWETHContract } from 'hooks/useContract'
 import { computeSlippageAdjustedAmounts } from 'utils/prices'
 import { PostOrderParams } from 'utils/trade'
