+++ src/cow-react/swap/services/swapFlow/index.ts
@@ -2,10 +2,10 @@ import { SwapFlowContext } from './types'
 import { swapFlowAnalytics } from './steps/analytics'
 import { signAndPostOrder } from 'utils/trade'
 import { presignOrderStep } from './steps/presignOrderStep'
-import { addPendingOrderStep } from 'pages/Swap/swapFlow/steps/addPendingOrderStep'
-import confirmPriceImpactWithoutFee from '@src/components/swap/confirmPriceImpactWithoutFee'
-import { logSwapFlow } from 'pages/Swap/swapFlow/logger'
-import { getSwapErrorMessage } from 'pages/Swap/swapFlow/steps/swapErrorHelper'
+import { addPendingOrderStep } from 'cow-react/swap/services/swapFlow/steps/addPendingOrderStep'
+import confirmPriceImpactWithoutFee from 'components/swap/confirmPriceImpactWithoutFee'
+import { logSwapFlow } from 'cow-react/swap/services/swapFlow/logger'
+import { getSwapErrorMessage } from 'cow-react/swap/services/swapFlow/steps/swapErrorHelper'
 import { PriceImpact } from 'hooks/usePriceImpact'
 
 export async function swapFlow(input: SwapFlowContext, priceImpactParams: PriceImpact) {
