+++ src/cow-react/swap/services/swapFlow/steps/addPendingOrderStep.ts
@@ -1,4 +1,4 @@
-import { serializeToken } from '@src/state/user/hooks'
+import { serializeToken } from 'state/user/hooks'
 import { addPendingOrder, AddPendingOrderParams, SerializedOrder } from 'state/orders/actions'
 import { AddUnserialisedPendingOrderParams } from 'state/orders/hooks'
 import { AppDispatch } from 'state'
