+++ src/cow-react/swap/services/swapFlow/steps/analytics.ts
@@ -1,6 +1,6 @@
 import { signSwapAnalytics, swapAnalytics } from 'utils/analytics'
 import TradeGp from 'state/swap/TradeGp'
-import { USER_SWAP_REJECTED_ERROR } from 'pages/Swap/swapFlow/steps/swapErrorHelper'
+import { USER_SWAP_REJECTED_ERROR } from 'cow-react/swap/services/swapFlow/steps/swapErrorHelper'
 
 function getMarketLabel(trade: TradeGp): string {
   return [trade.inputAmount.currency.symbol, trade.outputAmount.currency.symbol].join(',')
